API

Types

Ephemerides.AbstractEphemRecord — Type

AbstractEphemRecord

Abstract type for ephemeris segment records.

Ephemerides.AbstractSPKCache — Type

AbstractSPKCache

Abstract type for all SPK segment type caches.

Ephemerides.AbstractSPKHeader — Type

AbstractSPKHeader

Abstract type for all SPK segment type headers.

Ephemerides.AbstractSPKSegment — Type

AbstractSPKSegment

Abstract type for all SPK segment types.

Ephemerides.DAF — Type

DAF

Container to hold the information of NAIF's Double precision Array File (DAF).

Fields

filepath – String system filepath of the DAF

array – Vector{UInt8} binary content of the DAF

header – DAFHeader file record of the DAF

comment – String text within the DAF comment area

ftype – Int file type, equals 1 for SPK and 2 for PCK

desc – DAF PCK/SPK segment descriptors

seglist – SPKSegmentList list of the SPK/PCK segments within the DAF

References

DAF Required Reading

See Also

See also DAFHeader, SPKSegmentList and EphemerisProvider

Ephemerides.DAF — Method

DAF(filename::String)

Parse a DAF file and retrieve its content.

Note

This function does not initialise the segment types. That operation can be done in a second moment through the initialise_segments! function.

Warning

Only DAF files with identification word equal to "DAF/SPK" or "DAF/PCK" are accepted, otherwise an error is thrown.

See Also

See also initialise_segments!.

Ephemerides.DAFHeader — Type

DAFHeader

The DAF header, or file record, is the first physical record in a DAF and stores general information about the content of the file.

Fields

nd – Int32 number of double components in each array summary

ni – Int32 number of integer components in each array summary

fwd – Int32 record number of initial summary record

bwd – Int32 record number of final summary record

ffa – Int32 first free address of the file

name – String internal name of the file

lend – Bool true if the file was generated in little endian

References

DAF Required Reading

See Also

See also DAF and EphemerisProvider

Ephemerides.DAFHeader — Method

DAFHeader(record::Vector{UInt8})

Parse the header (file record) of the DAF file, i.e., the first physical record in a DAF which contains global information about the file.

References

DAF Required Reading

See Also

See also DAF, parse_daf_comment and parse_daf_summaries

Ephemerides.DAFSegmentDescriptor — Type

DAFSegmentDescriptor

A container object to store both SPK and PCK descriptors information.

Fields

segtype – Int32 SPK/PCK segment type

tstart – Float64 initial segment type, in TDB seconds since J2000.0

tend – Float64 final segment type, in TDB seconds since J2000.0

tid – Int32 target object NAIF ID

cid – Int32 center object NAIF ID

axesid – Int32 reference axes ID. Defaults to -1 for PCKs

iaa – Int32 initial array address

faa – Int32 final array address

References

SPK Required Reading

PCK Required Reading

Ephemerides.DAFSegmentDescriptor — Method

DAFSegmentDescriptor(summary::Vector{UInt8}, head::DAFHeader, isspk::Bool)

Generate an SPK or PCK descriptor by parsing a DAF summary.

Ephemerides.EphemRecordPCK — Type

EphemRecordPCK <: AbstractEphemRecord

Store the PCK metadata relative to a given (target, center) axes pair.

Fields

target – Int target axes NAIF ID

center – Int center axes NAIF ID

t_start – start times of each sub-window, in TDB seconds since J2000

t_end – final times of each sub-window, in TDB seconds since J2000

Ephemerides.EphemRecordSPK — Type

EphemRecordSPK <: AbstractEphemRecord

Store the SPK metadata relative to a given (target, center) objects pair.

Fields

target – Int target object NAIF ID

center – Int center object NAIF ID

axes – Int reference axes ID

t_start – start times of each sub-window, in TDB seconds since J2000

t_end – final times of each sub-window, in TDB seconds since J2000

Ephemerides.EphemerisProvider — Type

EphemerisProvider(file::String)
EphemerisProvider(files::Vector{String})

Create an EphemerisProvider instance by loading a single or multiple binary ephemeris kernel files specified by files. Currently, only NAIF Double precision Array File (DAF) kernels (i.e., SPK and PCK) are accepted.

Example

julia> eph = EphemerisProvider("PATH_TO_KERNEL")
EphemerisProvider([...])

julia> eph = EphemerisProvider(["PATH_TO_KERNEL_1", "PATH_TO_KERNEL_2"])
EphemerisProvider([])

Ephemerides.InterpCache — Type

InterpCache{T}

Ephemerides.InterpCache — Method

InterpCache{T}(len::Int, buffsize::Int)

Create an InterpCache instance of type T with len buffers, each of size buffsize.

Ephemerides.SPKLink — Type

SPKLink

A link object to create a mapping between DAFSegmentDescriptor and its actual location within an EphemerisProvider object.

Fields

desc – DAFSegmentDescriptor for the segment associated to this link

fid – Int index of the DAF containg the link data.

lid – Int field number in the SPKSegmentList for this segment type.

eid – Int index of the inner segment list that stores this SPK segment.

fct – Int 1 or -1 depending on whether the (from, to) directions must be reversed.

See Also

See also SPKLinkTable, SPKSegmentList and add_spklinks!.

Ephemerides.SPKLinkTable — Type

SPKLinkTable

Dictionary object providing all the SPKLink available between a set of (from, to) objects

Ephemerides.SPKSegmentCache1 — Type

SPKSegmentCache1 <: AbstractSPKCache

Cache instance for SPK segments of type 1 and 21. The fields contained within this cache are taken from the FORTRAN NAIF's SPICE implementation for type 1 SPK segments.

Ephemerides.SPKSegmentCache1 — Method

SPKSegmentCache1(head::SPKSegmentHeader1)

Initialise the cache for an SPK segment of type 1.

Ephemerides.SPKSegmentCache18 — Type

SPKSegmentCache18 <: AbstractSPKCache

Cache instance for SPK segments of type 18.

Ephemerides.SPKSegmentCache19 — Type

SPKSegmentCache19 <: AbstractSPKCache

Cache instance for SPK segments of type 19.

Ephemerides.SPKSegmentCache19 — Method

SPKSegmentCache19(daf::DAF, head::SPKSegmentHeader2)

Initialise the cache for an SPK segment of type 19.

Ephemerides.SPKSegmentCache2 — Type

SPKSegmentCache2 <: AbstractSPKCache

Cache instance for SPK segments of type 2 and 3.

Fields

A – Chebyshev's polynomial coefficients, with size (ncomp, order)

p – Stores the record mid point and radius and scale factor

buff – Stores the buffers for the Chebyshev polynomials

id – Index of the currently loaded logical record

Ephemerides.SPKSegmentCache2 — Method

SPKSegmentCache2(head::SPKSegmentHeader2)

Initialise the cache for an SPK segment of type 2 and 3.

Ephemerides.SPKSegmentCache20 — Type

SPKSegmentCache2 <: AbstractSPKCache

Cache instance for SPK segments of type 20.

Fields

A – Chebyshev's polynomial coefficients, with size (ncomp, order)

p – Stores the record mid point and radius and scale factor

x1 – Values of the Chebyshev's polynomials

x2 – Derivatives of the Chebyshev's polynomials

id – Index of the currently loaded logical record

Ephemerides.SPKSegmentCache20 — Method

SPKSegmentCache20(head::SPKSegmentHeader2)

Initialise the cache for an SPK segment of type 20.

Ephemerides.SPKSegmentCache8 — Type

SPKSegmentCache8 <: AbstractSPKCache

Cache instance for SPK segments of type 8 and 12.

Ephemerides.SPKSegmentCache8 — Method

SPKSegmentCache8(head::SPKSegmentHeader2)

Initialise the cache for an SPK segment of type 8 and 12.

Ephemerides.SPKSegmentCache9 — Type

SPKSegmentCache9 <: AbstractSPKCache

Cache instance for SPK segments of type 9 and 13.

Ephemerides.SPKSegmentCache9 — Method

SPKSegmentCache9(head::SPKSegmentHeader2)

Initialise the cache for an SPK segment of type 9 and 13.

Ephemerides.SPKSegmentHeader1 — Type

SPKSegmentHeader1 <: AbstractSPKHeader

Header instance for SPK segments of type 1 and 21.

Fields

n – Int number of records in the segment

ndirs – Int number of directory epochs

epochs – Storage for directory epochs or epochs (when ndirs = 0)

iaa - Int initial segment file address

etid – Int initial address for the epoch table (after all the MDA records)

recsize - Int Number of double numbers stored in each MDA record

maxdim - Int MDA dimension (fixed to 15 for type 1)

Ephemerides.SPKSegmentHeader1 — Method

SPKSegmentHeader1(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 1.

Ephemerides.SPKSegmentHeader18 — Type

SPKSegmentHeader18 <: AbstractSPKHeader

Header instance for SPK segments of type 18.

Fields

n – Int number of states in the segment

ndirs – Int number of epoch directories

epochs – Storage for directory epochs or epochs (when ndirs = 0)

iaa - Int initial segment file address

etid – Int initial address for the epoch table (after all the state data)

order – Int interpolating polynomial degree

N – Int group size

subtype – Int type 18 subtype, either 0 (Hermite) or 1 (Lagrange)

packetsize – Int packet size for each point, either 12 (Hermite) or 6 (Lagrange)

Ephemerides.SPKSegmentHeader18 — Method

SPKSegmentHeader18(n::Int)

Create a default header for type 18 segments, whose epoch array has length n.

Ephemerides.SPKSegmentHeader19 — Type

SPKSegmentHeader19 <: AbstractSPKHeader

Header instance for SPK segments of type 19.

Fields

n – Int number of states in the segment.

ndirs – Int number of epoch directories.

times – Storage for interval directories or start times (when ndirs = 0).

iaa - Int initial segment file address.

etid – Int byte address for the interval table (after all the minisegment data).

ptid – Int byte for the pointer table.

usefirst – Bool boundary flag, true if the preceding segment should be used.

type – Int either type 18 or 19.

Ephemerides.SPKSegmentHeader19 — Method

SPKSegmentHeader19(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 18 and 19.

Ephemerides.SPKSegmentHeader2 — Type

SPKSegmentHeader2 <: AbstractSPKHeader

Header instance for SPK segments of type 2 and 3.

Fields

tstart – Float64 initial epoch of the first record, in seconds since J2000

tlen – Float64 interval length covered by each record, in seconds

order – Int polynomial order

N – Int number of coefficients in each window

n – Int number of records in the segment

recsize – Int byte size of each logical record

ncomp – Int number of vector components

iaa – Int initial segment file address

type – Int SPK segment type, either 2 or 2

Ephemerides.SPKSegmentHeader2 — Method

SPKSegmentHeader2(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 2 and 3.

Ephemerides.SPKSegmentHeader20 — Type

SPKSegmentHeader20 <: AbstractSPKHeader

Header instance for SPK segments of type 20.

Fields

Ephemerides.SPKSegmentHeader20 — Method

SPKSegmentHeader20(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 20

Ephemerides.SPKSegmentHeader8 — Type

SPKSegmentHeader8 <: AbstractSPKHeader

Header instance for SPK segments of type 8 and 12.

Fields

tstart – Float64 segment starting epoch, in TDB seconds since J2000

tlen – Float64 interval length, in seconds

order – Int interpolating polynomial degree

N – Int group size

n – Int number of states in the segment

iaa - Int initial segment file address

iseven – Bool true for even group size

type – Int SPK type (either 8 or 12)

Ephemerides.SPKSegmentHeader8 — Method

SPKSegmentHeader8(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 8 and 12.

Ephemerides.SPKSegmentHeader9 — Type

SPKSegmentHeader9 <: AbstractSPKHeader

Header instance for SPK segments of type 9 and 13.

Fields

n – Int number of states in the segment

ndirs – Int number of epoch directories

epochs – Storage for directory epochs or epochs (when ndirs = 0)

iaa - Int initial segment file address

etid – Int initial address for the epoch table (after all the state data)

order – Int interpolating polynomial degree

N – Int group size

iseven – Bool true for even group size

type – Int SPK type (either 9 or 13)

Ephemerides.SPKSegmentHeader9 — Method

SPKSegmentHeader9(daf::DAF, desc::DAFSegmentDescriptor)

Create the segment header for an SPK segment of type 9 and 13.

Ephemerides.SPKSegmentList — Type

SPKSegmentList

A container object to efficiently store all the different SPK segments that are contained within a single DAF file.

SPKSegmentList()

Initialises an empty SPKSegmentList object.

See also

See also add_segment!

Ephemerides.SPKSegmentType1 — Type

SPKSegmentType1 <: AbstractSPKSegment

Segment instance for SPK segments of type 1 and 21, which contain Modified Difference Arrays (MDA). This data type is normally used for spacecraft whose ephemerides are produced by JPL's principal trajectory integrator DPTRAJ.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType1 — Method

SPKSegmentType1(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 1.

Ephemerides.SPKSegmentType19 — Type

SPKSegmentType19 <: AbstractSPKSegment

Segment instance for SPK segments of type 18 and 19. Type 18 segments are treated as special cases of a type 19 with a single mini-segment.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType19 — Method

SPKSegmentType19(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 19.

Ephemerides.SPKSegmentType2 — Type

SPKSegmentType2 <: AbstractSPKSegment

Segment instance for SPK segments of type 2 and 3, which contain Chebyshev polynomial coefficients for the position and/or state of the body as function of time. This data type is normally used for planet barycenters, and for satellites whose ephemerides are integrated.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType2 — Method

SPKSegmentType2(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 2.

Ephemerides.SPKSegmentType20 — Type

SPKSegmentType2 <: AbstractSPKSegment

Segment instance for SPK segments of type 20, which contain Chebyshev polynomial coefficients for the position and/or state of the body as function of time. This data type is normally used for planet barycenters, and for satellites whose ephemerides are integrated.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType20 — Method

SPKSegmentType20(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 2.

Ephemerides.SPKSegmentType8 — Type

SPKSegmentType8 <: AbstractSPKSegment

Segment instance for SPK segments of type 8 and 12.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType8 — Method

SPKSegmentType8(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 8 and 12.

Ephemerides.SPKSegmentType9 — Type

SPKSegmentType9 <: AbstractSPKSegment

Segment instance for SPK segments of type 9 and 13.

Fields

head – Segment header

cache – Segment cache

References

SPK Required Reading

SPICE Toolkit

Ephemerides.SPKSegmentType9 — Method

SPKSegmentType9(daf::DAF, desc::DAFSegmentDescriptor)

Create the object representing an SPK segment of type 9 and 13.

Functions

Ephemerides.add_segment! — Method

add_segment!(list::SPKSegmentList, spk::AbstractSPKSegment)

Add the SPK segment to the proper vector within the given SPKSegmentList list

Ephemerides.add_spklinks! — Method

add_spklinks!(table::SPKLinkTable, daf::DAF, fid::Int)

Insert in the input SPKLinkTable all the SPK or PCK links associated to the segment descriptors of the input DAF.

Ephemerides.analyse_timespan — Method

analyse_timespan(records)

Analyse a set of AbstractEphemRecord, returning the minimum and maximum covered times, in TDB seconds since J2000, together with a continuity parameter.

References

CALCEPH C++ library

See Also

See also ephem_spk_timespan and ephem_pck_timespan.

Ephemerides.array — Method

get_array(daf::DAF)

Return the byte content of the DAF file.

Ephemerides.axes — Method

axes(desc::DAFSegmentDescriptor)

Return the NAIF integer code for the reference axes. It is valid only for SPK files and defaults to -1 for PCKs.

Ephemerides.cache — Method

cache(spk::AbstractSPKSegment)

Return the segment cache data.

Ephemerides.center — Method

center(desc::DAFSegmentDescriptor)

Return the NAIF integer code for the reference object or axes for SPK and PCK, respectively.

Ephemerides.chebyshev — Function

chebyshev(cache::InterpCache, cₖ, t::Number, idx::Int, N::Int, ibuff::Int)

Evaluate a sum of Cheybyshev polynomials of the first kind at t using a recursive algorithm. It simultenously evalutes the 3 state components. idx is the index of the starting row (in 0-based notation) in the matrix of coefficients cₖ. ibuff is the index of the first free buffer.

Note

x is a re-work of the actual ascissa value that lies between [-1, 1]

See Also

See also ∂chebyshev, ∂²chebyshev and ∂³chebyshev

Ephemerides.comment — Method

get_comment(daf::DAF)

Return the comment written in the DAF comment section.

Ephemerides.compute_mda_position — Method

compute_mda_position(cache::SPKSegmentCache1, Δ::Number)

Ephemerides.compute_mda_velocity — Method

compute_mda_velocity(cache::SPKSegmentCache1, Δ::Number)

Ephemerides.create_linktables — Method

create_linktables(dafs::Vector{DAF})

Create the SPK and PCK SPKLinkTable for all the segments stores in the input DAFs.

Ephemerides.create_spk_segment — Method

create_spk_segment(daf::DAF, desc::DAFSegmentDescriptor)

Initialise an SPK segment according to the segment type defined in the DAFSegmentDescriptor desc.

Ephemerides.descriptor — Method

descriptor(link::SPKLink)

Return the SPK/PCK segment descriptor associated to this link.

Ephemerides.descriptors — Method

get_descriptors(daf::DAF)

Return the SPK/PCK segment descriptors contained in the DAF.

Ephemerides.element_id — Method

element_id(link::SPKLink)

Return the segment index in the inner SPK/PCK segment list.

Ephemerides.endian — Method

endian(head::DAFHeader)

Return true if the DAF file is in little-endian.

Ephemerides.endian — Method

endian(daf::DAF)

Return true if the DAF is in little-endian.

Ephemerides.ephem_get_axes — Method

ephem_get_axes(eph::EphemerisProvider)

Return a list of Frame IDs representing axes with available orientation data.

Ephemerides.ephem_get_points — Method

ephem_get_points(eph::EphemerisProvider)

Return a list of NAIFIds representing bodies with available ephemeris data.

Ephemerides.ephem_pck_records — Method

ephem_pck_records(eph::EphemerisProvider)

Return a list of EphemRecordPCK storing metadata relative to each (target, center) axes pairs in the loaded PCK kernels.

Ephemerides.ephem_pck_timespan — Method

ephem_pck_timespan(eph::EphemerisProvider)

Return the minimum and maximum time available in the PCK kernels loaded within eph, in TDB seconds since J2000, together with a continuity parameter defined as follows:

0 no PCK data is available.

1 the quantities of all axes are available for any time between the first and last time.

2 the quantities of some axes are available on discontinuous time intervals between the first and last time.

3 the quantities of each axis are available on a continuous time interval between the first and the last time, but not available for any time between the first and last time.

References

CALCEPH C++ library

See Also

See also ephem_spk_timespan.

Ephemerides.ephem_rotation12 — Method

ephem_rotation12(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 12-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation3 — Method

ephem_rotation3(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 3-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation6 — Method

ephem_rotation6(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 6-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_rotation9 — Method

ephem_rotation9(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 9-elements orientation angles of one set of axes (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Note

For the orientation angles, it is not possible to automatically compute the reverse transformation , i.e., if the orientation of PA440 is defined with respect to the ICRF, it is not possible to compute the rotation from the PA440 to the ICRF with this routine.

Ephemerides.ephem_spk_records — Method

ephem_spk_records(eph::EphemerisProvider)

Return a list of EphemRecordSPK storing metadata relative to each (target, center) object pairs in the loaded SPK kernels.

Ephemerides.ephem_spk_timespan — Method

ephem_spk_timespan(eph::EphemerisProvider)

Return the minimum and maximum time available in the SPK kernels loaded within eph, in TDB seconds since J2000, together with a continuity parameter defined as follows:

0 no SPK data is available.

1 the quantities of all bodies are available for any time between the first and last time.

2 the quantities of some bodies are available on discontinuous time intervals between the first and last time.

3 the quantities of each body are available on a continuous time interval between the first and the last time, but not available for any time between the first and last time.

References

CALCEPH C++ library

See Also

See also ephem_pck_timespan.

Ephemerides.ephem_timescale_id — Method

ephem_timescale_id(eph::EphemerisProvider)

Retrieve a timescale ID associated with the ephemeris handler eph. It returns 1 for Barycentric Dynamical Time (TDB) and 2 for Barycentric Coordinate Time (TCB).

Warning

Ephemeris providers with mixed timescales are not supported. An error is thrown if in the ephemeris handler some segments are defined in TDB and some other segments in TCB.

Ephemerides.ephem_vector12 — Method

ephem_vector12(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 12-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector3 — Method

ephem_vector3(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 3-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector6 — Method

ephem_vector6(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 6-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.ephem_vector9 — Method

ephem_vector9(eph::EphemerisProvider, from::Int, to::Int, time::Number)

Compute the 9-elements state of one body (to) relative to another (from) at time, expressed in TDB/TCB seconds since J2000, in accordance with the kernel timescale.

Ephemerides.factor — Method

factor(link::SPKLink)

Return the direction multiplicative factor.

Ephemerides.file_id — Method

file_id(link::SPKLink)

Return the DAF file index.

Ephemerides.filename — Method

filename(head::DAFHeader)

Return the internal description of the DAF.

Ephemerides.filepath — Method

filepath(daf::DAF)

Return the system path of the DAF.

Ephemerides.final_address — Method

final_address(desc::DAFSegmentDescriptor)

Return the final address of the segment array in teh DAF.

Ephemerides.final_record — Method

final_record(head::DAFHeader)

Return the record number of the final summary record in the DAF

Ephemerides.final_record — Method

final_record(daf::DAF)

Return the record number of the final summary record in the DAF.

Ephemerides.final_time — Method

final_time(desc::DAFSegmentDescriptor)

Return the final epoch of the interval for which ephemeris data are contained in the segment, in seconds since J2000.0

Ephemerides.final_time — Method

final_time(link::SPKLink)

Return the final epoch of the interval for which ephemeris data are contained in the segment associated to this link, in seconds since J2000.0

Ephemerides.final_times — Method

final_times(record::AbstractEphemRecord)

Recover the final times of each sub-window in which the ephemeris data of record is defined, expressed in TDB seconds since J2000

Ephemerides.find_logical_record — Method

find_logical_record(daf::DAF, head::SPKSegmentHeader1, time::Number)

Ephemerides.find_logical_record — Method

find_logical_record(daf::DAF, head::SPKSegmentHeader18, time::Number)

Ephemerides.find_logical_record — Method

find_logical_record(daf::DAF, head::SPKSegmentHeader1, time::Number)

Ephemerides.find_logical_record — Method

find_logical_record(head::SPKSegmentHeader2, time::Number)

Ephemerides.find_logical_record — Method

find_logical_record(head::SPKSegmentHeader2, time::Number)

Ephemerides.find_logical_record — Method

find_logical_record(head::SPKSegmentHeader8, time::Number)

Ephemerides.find_minisegment — Method

find_minirecord(daf::DAF, head::SPKSegmentHeader19, time::Number)

Ephemerides.free_address — Method

free_address(head::DAFHeader)

Return the first free address in the file, i.e., the address at which the first element of the next array is to be added.

Ephemerides.free_address — Method

free_address(daf::DAF)

Return the first free address in the file, i.e., the address at which the first element of the next array is to be added.

Ephemerides.get_buffer — Method

get_buffer(c::InterpCache, idx::int, x::Number)

Return the idx-th buffer from the corresponding DiffCache depending on the type of x.

Ephemerides.get_coefficients! — Method

get_coefficients!(daf::DAF, head::SPKSegmentHeader1, cache::SPKSegmentCache1, index::Int)

Ephemerides.get_coefficients! — Method

get_coefficients!(daf::DAF, head, cache, index::Int)

Ephemerides.get_coefficients! — Method

get_coefficients!(daf::DAF, head, cache, index::Int)

Ephemerides.get_coefficients! — Method

get_coefficients!(daf::DAF, head, cache, index::Int)

Ephemerides.get_daf — Method

daf(eph::EphemerisProvider, id::Int)

Return the DAF file in the ephemeris provider at index id.

Ephemerides.get_daf — Method

daf(eph::EphemerisProvider)

Return the DAF files stored in the ephemeris provider.

Ephemerides.get_record — Method

get_record(array::Vector{UInt8}, index::Integer)

Retrieve a whole DAF record at position index.

Ephemerides.get_segment — Method

get_segment(list::SPKSegmentList, lid::Int, eid::Int)

Return the segment contained in the lid list at index eid.

Ephemerides.get_segment_boundaries — Method

get_segment_boundaries(desclist::Vector{DAFSegmentDescriptor})

Parse all the segment descriptors of a given (center, target) pair and return a set of initial and final times, in TDB seconds since J2000, representing all the time sub-windows in which the ephemeris data for this pair is defined.

Ephemerides.header — Method

get_header(daf::DAF)

Return the DAFHeader header of the DAF.

Ephemerides.header — Method

header(spk::AbstractSPKSegment)

Return the segment header.

Ephemerides.hermite — Method

hermite(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Evalute a Hermite polynomial at x using a recursive algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.hermite — Method

hermite(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Evalute a Hermite polynomial at x using a recursive algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δt is the length of the polynomial interval.

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also ∂hermite, ∂²hermite and ∂³hermite.

Ephemerides.initial_address — Method

initial_address(desc::DAFSegmentDescriptor)

Return the initial address of the segment array in the DAF.

Ephemerides.initial_record — Method

initial_record(head::DAFHeader)

Return the record number of the initial summary record in the DAF

Ephemerides.initial_record — Method

initial_record(daf::DAF)

Return the record number of the initial summary record in the DAF.

Ephemerides.initial_time — Method

initial_time(desc::DAFSegmentDescriptor)

Return the initial epoch of the interval for which ephemeris data are contained in the segment, in seconds since J2000.0

Ephemerides.initial_time — Method

initial_time(link::SPKLink)

Return the initial epoch of the interval for which ephemeris data are contained in the segment associated to this link, in seconds since J2000.0

Ephemerides.initial_times — Method

initial_times(record::AbstractEphemRecord)

Recover the initial times of each sub-window in which the ephemeris data of record is defined, expressed in TDB seconds since J2000

Ephemerides.initialise_segments! — Method

initialise_segments!(daf::DAF)

Fill the SPKSegmentList by initialising the SPK/PCK segments associated to all the descriptors stores within the DAF.

See Also

See also DAF and create_spk_segment

Ephemerides.is_little_endian — Method

is_little_endian(array::Vector{UInt8})

Return true if the array corresponds to the string indicating a little-endian format.

Ephemerides.is_pck — Method

is_pck(daf::DAF)

Return true if the DAF stores PCK data.

Ephemerides.is_spk — Method

is_spk(daf::DAF)

Return true if the DAF stores SPK data.

Ephemerides.lagrange — Method

lagrange(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Recursively evaluate a Lagrange polynomial at x by using Neville's algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.lagrange — Method

lagrange(cache::InterpCache, states, x, idx::Int, N::Int)

Recursively evaluate a Lagrange polynomial at x by using Neville's algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state and N is the number of coefficients of the polynomial.

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also ∂lagrange and ∂²lagrange.

Ephemerides.list_id — Method

list_id(link::SPKLink)

Return the index of the list containing the segments of the given SPK/PCK type.

Ephemerides.normalise_time — Method

normalise_time(head::SPKSegmentHeader8, time::Number, index::Int)

Returned a normalised time that starts at 1 at the beginning of the interval.

Ephemerides.parse_daf_comment — Method

parse_daf_comment(array::Vector{UInt8}, header::DAFHeader)

Retrieve the comment section of a binary DAF.

References

DAF Required Reading

See Also

See also DAF, DAFHeader and parse_daf_summaries

Ephemerides.parse_daf_summaries — Method

parse_daf_summaries(array::Vector{UInt8}, head::DAFHeader)

Parse the DAF binary content and retrieve all the summary records.

References

DAF Required Reading

See Also

See also DAF, DAFHeader and parse_daf_comment

Ephemerides.parse_pck_segment_descriptor — Method

parse_pck_segment_descriptor(summary::Vector{UInt8}, lend::Bool)

Create a DAFSegmentDescriptor object by parsing a binary PCK segment descriptor. A default value of -1 is used to fill the reference frame field. The target and center fields are used for the actual target and center axes.

References

PCK Required Reading

Ephemerides.parse_spk_segment_descriptor — Method

parse_spk_segment_descriptor(summary::Vector{UInt8}, lend::Bool)

Create a DAFSegmentDescriptor object by parsing a binary SPK segment descriptor.

References

SPK Required Reading

Ephemerides.pck_links — Method

pck_links(eph::EphemerisProvider)

Return the SPKLinkTable for the PCK segments.

Ephemerides.reset_indexes! — Method

reset_indexes!(cache::SPKSegmentCache18)

Reset the cache indexes to force the coefficients reload.

Ephemerides.reverse_link — Method

reverse_link(link::SPKLink)

Reverse the sign, i.e. change the sign of the multiplicative factor, of the link.

Ephemerides.segment_list — Method

get_segment_list(daf::DAF)

Return the SPKSegmentList list of segments stored in the DAF.

Ephemerides.segment_type — Method

segment_type(desc::DAFSegmentDescriptor)

Return the SPK/PCK segment type.

Ephemerides.spk_field — Method

spk_field(spk::AbstractSPKSegment)

Return the field number in the SPKSegmentList associated to the given SPK segment type.

Ephemerides.spk_links — Method

spk_links(eph::EphemerisProvider)

Return the [SPKLinkTable] for the SPK segments.

Ephemerides.summary_size — Method

summary_size(head::DAFHeader)

Compute the size of a single summary record of a DAF file, in bytes.

Ephemerides.target — Method

target(desc::DAFSegmentDescriptor)

Return the NAIF integer code for the target object or axes for SPK and PCK, respectively.

Ephemerides.update_header! — Method

update_header!(head::SPKSegmentHeader18, daf::DAF, iaa, faa, type)

Update the header of type 18 segments.

Ephemerides.∂chebyshev — Function

∂chebyshev(cache::InterpCache, cₖ, t::Number, idx::Int, N::Int, Δt, ibuff=1)

Evaluate a sum of Cheybyshev polynomials of the first kind and its derivative at t using a recursive algorithm. It simultenously evalutes the 3 state components. idx is the index of the starting row (in 0-based notation) in the matrix of coefficients cₖ. ibuff is the index of the first free buffer.

Note

x is a re-work of the actual ascissa value that lies between [-1, 1]

See Also

See also chebyshev, ∂²chebyshev and ∂³chebyshev

Ephemerides.∂hermite — Method

∂hermite(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Evalute a Hermite polynomial and its derivative at x using a recursive algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.∂hermite — Method

∂hermite(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Evalute a Hermite polynomial and its derivative at x using a recursive algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δt is the length of the polynomial interval.

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also hermite, ∂²hermite and ∂³hermite.

Ephemerides.∂lagrange — Method

∂lagrange(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Recursively evaluate a Lagrange polynomial and its derivative at x by using Neville's algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.∂lagrange — Method

∂lagrange(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Recursively evaluate a Lagrange polynomial and its derivative at x by using Neville's algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δtis the length of the polynomial interval

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also lagrange and ∂²lagrange

Ephemerides.∂²chebyshev — Function

∂²chebyshev(cache::InterpCache, cₖ, t::Number, idx::Int, N::Int, Δt, ibuff=1)

Evaluate a sum of Cheybyshev polynomials of the first kind and its two derivatives at t using a recursive algorithm. It simultenously evalutes the 3 state components. idx is the index of the starting row (in 0-based notation) in the matrix of coefficients cₖ. ibuff is the index of the first free buffer.

Note

x is a re-work of the actual ascissa value that lies between [-1, 1]

See Also

See also chebyshev, ∂chebyshev and ∂³chebyshev

Ephemerides.∂²hermite — Method

∂²hermite(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Evalute a Hermite polynomial and its two derivatives at x using a recursive algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.∂²hermite — Method

∂²hermite(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Evalute a Hermite polynomial and its two derivatives at x using a recursive algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δt is the length of the polynomial interval.

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also hermite, ∂hermite and ∂³hermite.

Ephemerides.∂²lagrange — Method

∂²lagrange(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Recursively evaluate a Lagrange polynomial and its two derivatives at x by using Neville's algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

See Also

See also lagrange and ∂lagrange

Ephemerides.∂²lagrange — Method

∂²lagrange(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Recursively evaluate a Lagrange polynomial and its two derivatives at x by using Neville's algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δtis the length of the polynomial interval

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also lagrange and ∂lagrange

Ephemerides.∂³chebyshev — Function

∂³chebyshev(cache::InterpCache, cₖ, t::Number, idx::Int, N::Int, Δt, ibuff=1)

Evaluate a sum of Cheybyshev polynomials of the first kind and its three derivatives at t using a recursive algorithm. It simultenously evalutes the 3 state components. idx is the index of the starting row (in 0-based notation) in the matrix of coefficients cₖ. ibuff is the index of the first free buffer.

Note

x is a re-work of the actual ascissa value that lies between [-1, 1]

See Also

See also chebyshev, ∂chebyshev and ∂²chebyshev

Ephemerides.∂³hermite — Method

∂³hermite(cache::InterpCache, states, epochs, x, idx::Int, N::Int)

Evalute a Hermite polynomial and its three derivatives at x using a recursive algorithm. This function handles unequally-spaced polynomials, where the coefficients in states are interpolated at epochs.

Ephemerides.∂³hermite — Method

∂³hermite(cache::InterpCache, states, x, idx::Int, N::Int, Δt::Number)

Evalute a Hermite polynomial and its three derivatives at x using a recursive algorithm. This function is valid only for equally-spaced polynomials. idx is the index of the desired state, N is the number of coefficients of the polynomial and Δt is the length of the polynomial interval.

Note

x is a re-work of the actual ascissa value that starts at 1

See Also

See also hermite, ∂hermite and ∂²hermite.

Ephemerides.∫chebyshev — Method

∫chebyshev(cache::InterpCache, cₖ, t::Number, N::Int, Δt, tlen, p₀)

Evaluate the integral of a sum of Cheybyshev polynomials of the first kind using a recursive algorithm. It simultenously evalutes the 3 state components. This function simultaneously computes both the Chebyshev polynomials as well as their integrals.

Ephemerides.∫chebyshev — Method

∫chebyshev(cache::InterpCache, cₖ, t::Number, N::Int, Δt, tlen, p₀)

Evaluate the integral of a sum of Cheybyshev polynomials of the first kind using a recursive algorithm. It simultenously evalutes the 3 state components. It assumes the Chebyshev polynomials up to degree N have already been computed and are stored in the buffer with index ibuff. tlen is the size of the record interval, Δt is the timescale factor, and p₀ is a vector containing the position coefficients at the midpoint (i.e., when the integral is evaluated at t = 0).

Note

x is a re-work of the actual ascissa value that lies between [-1, 1]

JSMDInterfaces.Ephemeris.ephem_available_axes — Method

ephem_available_points(eph::EphemerisProvider)

Return a list of Frame IDs representing axes with available orientation data.

JSMDInterfaces.Ephemeris.ephem_available_points — Method

ephem_available_points(eph::EphemerisProvider)

Return a list of NAIFIds representing bodies with available ephemeris data.

JSMDInterfaces.Ephemeris.ephem_compute! — Method

ephem_compute!(res, eph, jd0, time, target, center, order)

Interpolate the position and/or its derivatives up to order for one body target relative to another center at the time jd0 + time, expressed as a Julian Date. This function reads the ephemeris files associated to eph and stores the results to res.

The returned array res must be large enough to store the results. The size of this array must be equal to 3*order:

res[1:3] contain the position (x, y, z) and is always valid

res[4:6] contain the velocity (dx/dt, dy/dt, dz/dt) for order ≥ 1

res[7:9] contain the acceleration (d²x/dt², d²y/dt², d²z/dt²) for order ≥ 2

res[10:12] contain the jerk (d³x/dt³, d³y/dt³, d³z/dt³) for order ≥ 3

The values stores in res are always returned in km, km/s, km/s², km/s³

JSMDInterfaces.Ephemeris.ephem_orient! — Method

ephem_orient!(res, eph, jd0, time, target, center, order)

Interpolate the orientation and its derivatives up to order for the target body at the time jd0 + time, expressed as a Julian Date. This function reads the ephemeris files associated to eph and stores the results to res.

The returned array res must be large enough to store the results. The size of this array must be equal to 3*order:

res[1:3] contain the angles

res[4:6] contain the 1st derivative for order ≥ 1

res[7:9] contain the 2nd derivative for order ≥ 2

res[10:12] contain the 3rd derivative for order ≥ 3

The values stores in res are always returned in rad, rad/s, rad/s², rad/s³

JSMDInterfaces.Ephemeris.ephem_orient_records — Method

ephem_orient_records(eph::EphemerisProvider)

Get an array of EphemAxesRecord, providing detailed informations on the orientation content of the ephemeris file.

JSMDInterfaces.Ephemeris.ephem_position_records — Method

ephem_position_records(eph::EphemerisProvider)

Get an array of EphemPointRecord, providing detailed informations on the position content of the ephemeris file.

JSMDInterfaces.Ephemeris.ephem_timescale — Method

ephem_timescale(eph::EphemerisProvider)

Retrieve a timescale ID associated with the ephemeris handler eph. It returns 1 for Barycentric Dynamical Time (TDB) and 2 for Barycentric Coordinate Time (TCB).

Warning

An error is thrown if the timescale is neither TDB nor TCB.

JSMDInterfaces.Ephemeris.ephem_timespan — Method

ephem_timespan(eph::EphemerisProvider)

Returns the first and last time available in the ephemeris file associated to an ephemeris file. It returns a tuple containing:

firsttime – Julian date of the first time.

lasttime – Julian date of the last time.

continuous – Information about the availability of the quantities over the time span. It equals:

1: if the quantities of all bodies are available for any time between the first and last time.

2: if the quantities of some bodies are available on discontinuous time intervals between the first and last time.

3: if the quantities of each body are available on a continuous time interval between the first and last time, but not available for any time between the first and last time.

Constants

Ephemerides.DAF_RECORD_LENGTH — Constant

DAF_RECORD_LENGTH

DAF record length, in bytes.

References

DAF Required Reading

Ephemerides.FTPSTR — Constant

FTPSTR

Validation string that guarantees the integrity of a DAF file.

References

DAF Required Reading

Ephemerides.SPK_SEGMENTLIST_MAPPING — Constant

SPK_SEGMENT_MAPPING

A dictionary mapping SPK segment types to the field index of the SPKSegmentList.

Ephemerides.TCB_SEGMENTS — Constant

TCB_SEGMENTS

List of the SPK/PCK segment types for which the time argument is expressed in the TCB scale.

Ephemerides.TDB_SEGMENTS — Constant

TDB_SEGMENTS

List of the SPK/PCK segment types for which the time argument is expressed in the TDB scale.